* Get rid of custom min/max macros * Fixed some compiler warnings * Fixed compilation on MacOS X 10.6 * Code cleanups * Report nick!ident@host for some events (not quite final yet) * Fixed a few memory leaks * Added missing GPL header for src/sbnc.h * Use system-provided snprintf() when available
git-svn-id: http://svn.beutner.name/trunk/sbnc@1103 b3a6972a-4407-0410-a3f0-9bc6c1699d69

+++ src/FloodControl.cpp
@@ -145,7 +145,7 @@ RESULT<char *> CFloodControl::DequeueItem(bool Peek) {
 	THROWIFERROR(char *, Item);
 
 	if (m_Control) {
-		m_Bytes = GetBytes() + max(130, strlen(Item) * CalculatePenaltyAmplifier(Item));
+		m_Bytes = GetBytes() + max((size_t)130, strlen(Item) * CalculatePenaltyAmplifier(Item));
 
 		ScheduleItem();
 	}
